+++ src/zenml/orchestrators/step_run_utils.py
@@ -20,7 +20,8 @@ from zenml.client import Client
 from zenml.config.step_configurations import Step
 from zenml.constants import CODE_HASH_PARAMETER_NAME, TEXT_FIELD_MAX_LENGTH
 from zenml.enums import ExecutionStatus
-from zenml.execution.step_runtime import BaseStepRuntime, DefaultStepRuntime
+from zenml.execution.default_runtime import DefaultStepRuntime
+from zenml.execution.step_runtime import BaseStepRuntime
 from zenml.logger import get_logger
 from zenml.model.utils import link_artifact_version_to_model_version
 from zenml.models import (
