+++ tests/unit/execution/test_default_runtime_metadata_toggle.py
@@ -2,19 +2,22 @@
 
 from types import SimpleNamespace
 
-from zenml.execution.step_runtime import DefaultStepRuntime
+from zenml.execution.default_runtime import DefaultStepRuntime
 
 
 def test_publish_metadata_skips_when_disabled(monkeypatch):
+    """Test that metadata is not published when disabled."""
     rt = DefaultStepRuntime()
     setattr(rt, "_metadata_enabled", False)
 
     called = {"run": 0, "step": 0}
 
     def _pub_run_md(*a, **k):
+        """Mock publish pipeline run metadata."""
         called["run"] += 1
 
     def _pub_step_md(*a, **k):
+        """Mock publish step run metadata."""
         called["step"] += 1
 
     monkeypatch.setattr(
